IGNITE-22141 moved job to correct class
