+++ modules/runner/src/testFixtures/java/org/apache/ignite/internal/runner/app/Jobs.java
@@ -30,6 +30,7 @@ import java.math.BigDecimal;
 import java.math.RoundingMode;
 import java.util.ArrayList;
 import java.util.Arrays;
+import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.Set;
@@ -47,6 +48,8 @@ import org.apache.ignite.marshalling.ByteArrayMarshaller;
 import org.apache.ignite.marshalling.Marshaller;
 import org.apache.ignite.network.ClusterNode;
 import org.apache.ignite.table.Tuple;
+import org.apache.ignite.table.partition.Partition;
+import org.apache.ignite.table.partition.PartitionDistribution;
 import org.jetbrains.annotations.Nullable;
 
 /** Jobs and marshallers definitions that are used in tests. */
@@ -628,6 +631,24 @@ public class Jobs {
         @Override
         public CompletableFuture<String> reduceAsync(TaskExecutionContext context, Map<UUID, String> results) {
             throw new CustomException(TRACE_ID, COLUMN_NOT_FOUND_ERR, "Custom job error", null);
+        }
+    }
+    /**
+     * Job returns actual partition distribution for particular table calculated by java side.
+     */
+    public static class GetPartitionDistributionByTableJob implements ComputeJob<String, Map<UUID, List<Long>>> {
+        @Override
+        public @Nullable CompletableFuture<Map<UUID, List<Long>>> executeAsync(JobExecutionContext context, String tableName) {
+            PartitionDistribution pd = context.ignite().tables().table(tableName).partitionDistribution();
+
+            Map<UUID, List<Long>> distribution = new HashMap<>();
+            for (Partition partition : pd.partitions()) {
+                var primaryNode = pd.primaryReplica(partition);
+
+                distribution.computeIfAbsent(primaryNode.id(), k -> new ArrayList<>()).add(partition.id());
+            }
+
+            return completedFuture(distribution);
         }
     }
 }
